+++ static/app/views/issueDetails/streamline/header.spec.tsx
@@ -2,13 +2,11 @@ import {GroupFixture} from 'sentry-fixture/group';
 import {LocationFixture} from 'sentry-fixture/locationFixture';
 import {OrganizationFixture} from 'sentry-fixture/organization';
 import {ProjectFixture} from 'sentry-fixture/project';
-import {ReleaseFixture} from 'sentry-fixture/release';
 import {RouterFixture} from 'sentry-fixture/routerFixture';
 import {TeamFixture} from 'sentry-fixture/team';
 import {UserFixture} from 'sentry-fixture/user';
 
 import {render, screen} from 'sentry-test/reactTestingLibrary';
-import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import type {TeamParticipant, UserParticipant} from 'sentry/types/group';
 import {IssueCategory} from 'sentry/types/group';
@@ -45,9 +43,6 @@ describe('UpdatedGroupHeader', () => {
       project,
     };
 
-    const firstRelease = ReleaseFixture({id: '1'});
-    const lastRelease = ReleaseFixture({id: '2'});
-
     beforeEach(() => {
       MockApiClient.addMockResponse({
         url: '/organizations/org-slug/replay-count/',
@@ -58,14 +53,7 @@ describe('UpdatedGroupHeader', () => {
         url: `/organizations/org-slug/repos/`,
         body: {},
       });
-      MockApiClient.addMockResponse({
-        url: `/projects/org-slug/project-slug/releases/${encodeURIComponent(firstRelease.version)}/`,
-        body: {},
-      });
-      MockApiClient.addMockResponse({
-        url: `/organizations/org-slug/releases/${encodeURIComponent(firstRelease.version)}/deploys/`,
-        body: {},
-      });
+
       MockApiClient.addMockResponse({
         url: `/organizations/${organization.slug}/issues/${group.id}/attachments/`,
         body: [],
@@ -77,11 +65,6 @@ describe('UpdatedGroupHeader', () => {
     });
 
     it('shows all elements of header', async () => {
-      MockApiClient.addMockResponse({
-        url: `/organizations/${organization.slug}/issues/${group.id}/first-last-release/`,
-        method: 'GET',
-        body: {firstRelease, lastRelease},
-      });
       const teams: TeamParticipant[] = [{...TeamFixture(), type: 'team'}];
       const users: UserParticipant[] = [
         {
@@ -133,11 +116,6 @@ describe('UpdatedGroupHeader', () => {
       expect(
         await screen.findByRole('link', {name: formatAbbreviatedNumber(group.userCount)})
       ).toBeInTheDocument();
-
-      expect(
-        await screen.findByText(textWithMarkupMatcher('Releases'))
-      ).toBeInTheDocument();
-
       expect(
         screen.getByRole('button', {name: 'Modify issue priority'})
       ).toBeInTheDocument();
@@ -155,36 +133,7 @@ describe('UpdatedGroupHeader', () => {
       expect(screen.getByRole('button', {name: 'Archive'})).toBeInTheDocument();
     });
 
-    it('only shows one release if possible', async function () {
-      MockApiClient.addMockResponse({
-        url: `/organizations/${organization.slug}/issues/${group.id}/first-last-release/`,
-        method: 'GET',
-        // First and last release match
-        body: {firstRelease, lastRelease: firstRelease},
-      });
-      render(
-        <StreamlinedGroupHeader
-          {...defaultProps}
-          group={group}
-          project={project}
-          event={null}
-        />,
-        {
-          organization,
-          router,
-        }
-      );
-      expect(
-        await screen.findByText(textWithMarkupMatcher('Release'))
-      ).toBeInTheDocument();
-    });
-
     it('displays new experience button if flag is set', async () => {
-      MockApiClient.addMockResponse({
-        url: `/organizations/${organization.slug}/issues/${group.id}/first-last-release/`,
-        method: 'GET',
-        body: {firstRelease, lastRelease},
-      });
       const flaggedOrganization = OrganizationFixture({
         features: ['issue-details-streamline'],
       });
